Clean up the account_review_v2 feature flag
