+++ clients/apps/web/src/app/(main)/dashboard/[organization]/(header)/finance/account/AccountPageDetailsRequired.tsx
@@ -1,36 +1,119 @@
 'use client'
 
+import { ReviewChecklist } from '@/components/Finance/Account/ReviewChecklist'
 import { DashboardBody } from '@/components/Layout/DashboardLayout'
-import OrganizationProfileSettings from '@/components/Settings/OrganizationProfileSettings'
-import { Section, SectionDescription } from '@/components/Settings/Section'
-import { useHasPermission } from '@/hooks/permissions'
+import { toast } from '@/components/Toast/use-toast'
+import { usePostHog } from '@/hooks/posthog'
+import { useProducts } from '@/hooks/queries'
+import {
+  useOrganizationKYC,
+  useOrganizationReviewState,
+  useSubmitOrganizationReview,
+} from '@/hooks/queries/org'
+import {
+  usePayoutAccount,
+  usePayoutAccounts,
+} from '@/hooks/queries/payout_accounts'
 import { schemas } from '@polar-sh/client'
+import { Text } from '@polar-sh/orbit'
+import { Box } from '@polar-sh/orbit/Box'
+import { Button } from '@polar-sh/orbit'
+import { useRouter } from 'next/navigation'
+import { useState } from 'react'
+
+const ROW_STAGGER = 0.06
+const ROW_DURATION = 0.32
 
 interface Props {
   organization: schemas['Organization']
 }
 
 export const AccountPageDetailsRequired = ({ organization }: Props) => {
-  const canManageOrganization = useHasPermission(
+  const router = useRouter()
+  const { data: reviewState, isLoading } = useOrganizationReviewState(
     organization.id,
-    'organization:manage',
   )
 
+  // Warm up query caches to prevent flickering
+  useOrganizationKYC(organization.id)
+  useProducts(organization.id, { limit: 1 })
+  usePayoutAccounts()
+  usePayoutAccount(organization.payout_account_id || undefined)
+
+  const submitReview = useSubmitOrganizationReview(organization.id)
+  const posthog = usePostHog()
+
+  const [isExiting, setIsExiting] = useState(false)
+
+  const steps = reviewState?.preliminary_steps ?? []
+  const rowsExitMs =
+    (Math.max(0, steps.length - 1) * ROW_STAGGER + ROW_DURATION) * 1000
+
+  const handleSubmit = async () => {
+    if (isExiting) return
+
+    posthog.capture('dashboard:organizations:account_review:submit', {
+      organization_id: organization.id,
+      section: 'cta',
+    })
+
+    setIsExiting(true)
+
+    const [{ error }] = await Promise.all([
+      submitReview.mutateAsync(),
+      new Promise((resolve) => setTimeout(resolve, rowsExitMs)),
+    ])
+
+    if (error) {
+      toast({
+        title: 'Submission failed',
+        description:
+          'We could not submit your organization for review. Please try again.',
+      })
+      setIsExiting(false)
+      return
+    }
+
+    posthog.capture('dashboard:organizations:account_review:done', {
+      organization_id: organization.id,
+      section: 'cta',
+    })
+
+    router.refresh()
+  }
+
   return (
-    <DashboardBody wrapperClassName="max-w-(--breakpoint-sm)!">
-      <div className="flex flex-col gap-y-12">
-        <Section>
-          <SectionDescription
-            title="Account Review"
-            description="Tell us about your organization so we can review your usecase."
-          />
-          <OrganizationProfileSettings
-            organization={organization}
-            kyc={true}
-            readOnly={!canManageOrganization}
-          />
-        </Section>
-      </div>
+    <DashboardBody
+      wrapperClassName="max-w-(--breakpoint-sm)!"
+      title="Account Review"
+      header={
+        <Button
+          onClick={handleSubmit}
+          disabled={!reviewState?.can_submit || isExiting}
+          loading={isExiting}
+        >
+          Submit for review
+        </Button>
+      }
+    >
+      <Box
+        display="flex"
+        flexDirection="column"
+        rowGap="xl"
+        paddingBottom="3xl"
+      >
+        <Text variant="body" color="muted">
+          Verify your business so customers can buy from you. After you submit,
+          our team will review your details and get back to you shortly.
+        </Text>
+        <ReviewChecklist
+          isLoading={isLoading}
+          isExiting={isExiting}
+          steps={steps}
+          rowStagger={ROW_STAGGER}
+          rowDuration={ROW_DURATION}
+        />
+      </Box>
     </DashboardBody>
   )
 }
