+++ clients/apps/web/src/app/(main)/dashboard/[organization]/(header)/finance/account/AccountPageRouter.tsx
@@ -4,7 +4,6 @@ import { useOrganizationReviewStatus } from '@/hooks/queries/org'
 import { schemas } from '@polar-sh/client'
 import { AccountPageApproved } from './AccountPageApproved'
 import { AccountPageDetailsRequired } from './AccountPageDetailsRequired'
-import { AccountPageDetailsRequiredV2 } from './AccountPageDetailsRequiredV2'
 import { AccountPageInReview } from './AccountPageInReview'
 
 interface Props {
@@ -41,9 +40,6 @@ export const AccountPageRouter = ({
       isActive
 
   if (requireDetails) {
-    if (organization.feature_settings?.account_review_v2_enabled) {
-      return <AccountPageDetailsRequiredV2 organization={organization} />
-    }
     return <AccountPageDetailsRequired organization={organization} />
   }
